Improving DB setup
